requirements.txt
flake8
pycodestyle
mypy
ruff
pylint
pytest
allure-pytest
selenium
requests
playwright